feat : Equipment Bonus Stat
장비 장착시 보너스 스탯에 장비의 스텟 적용 로직 추가

+++ Assets/1. Scripts/Interactable/Item/ItemData.cs
@@ -25,8 +25,8 @@ public class Buff
     [Header("Buff Info")]
     [SerializeField] BuffType type;
     public BuffType Type => type;
-    [SerializeField] float value;
-    public float Value => value;
+    [SerializeField] int value;
+    public int Value => value;
 
     [Header("Duration")]
     [SerializeField] bool hasDuration;
